+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.0.0-pre.7] - 2020-07-19
+
+### Added
+
+- `StringView#includes` method.
+- `ArrayView#at` & `VectorView#at` methods.
+
+### Changed
+
+- `StringView#search` renamed into `StringView#indexof`.
+
 ## [4.0.0-pre.6] - 2020-06-26
 
 ### Added